programming-phoenix: test User model
First part of "Splitting Side Effects in Model Tests" section.

This was tricky to get right! The trickiest part was that the book was
outdated with the first `registration_changeset` test. I didn't fully
understand how the `in` tests were working until I played around in
`iex` for a while, and figured out that the reason it isn't working is
because `count: 6` isn't the only tuple in the keyword list anymore. And
there's no way to use pattern matching with `in`.

So I split that test into a couple lines, rather than putting it all in
one.